Streamlined error handling with a HttpError class.
This avoids all of the try/catch ugliness. In addition, we now only throw
HttpErrors in the handlers, and rely on the catch at the top-level listener to
convert these into errorResponse. This avoids the need to choose between
errorResponse and HttpError throughout.

To ensure that all non-blocking cache.put jobs are properly waited upon, we
provide a top-level array that collects all non-blocking jobs throughout the
handler. The listener is then responsible for calling event.waitUntil. This
ensures that we wait on the non-blocking jobs, even if an error is thrown.

Also, we streamlined the process of cache.put by adding a utility to handle
all the Response + Content-Type boilerplate.