Simplify the handlers to avoid implicit use of global variables.
- Explicitly pass the R2 bucket bindings to each handler.
- Centralize all global environment variables into a single object.

Also, wait for the handler's response to resolve before returning the
non-blocking array, otherwise the array might not even be filled yet.

On an unrelated note, we also force a reload from the bucket when updating the
permissions, so as to ensure that repeated calls will not accidentally use the
values in the cache.