Added MD5 integrity checks to all presigned URLs. (#7)
Also return arrays instead of objects in the upload start response. This is
more consistent with other response formats and allows us to re-use the
base64-encoded MD5 sum in the client's PUT request. (Why the presigned
URLs need the MD5 sum again is a mystery, given it's in the URL itself!)