Allow GitHub organizations to be listed in permissions.
Specifically, permissions checking will consider the user to belong to a
permissions list if any of its orgs are in that list. findUser now
returns an object containing the login and the orgs to which it belongs.
This requires some refactoring to distinguish between of downstream
functions that require the user_name and those that require the user
information object returned by findUser.