Update to a major version (v3) of the AWS SDK. (#10)
This is necessary because Wrangler 2 (and specifically esbuild) does not like
aspects of the SDK v2 during bundling, which breaks dev/publish.  On the other
hand - jest does not like aspects of the SDK v3, specifically relating to the
uuid package, and this requires some workarounds until jest/uuid sort it out.