Added an override key to allow uploads from any user.
This avoids the need to continually update ALLOWED_UPLOADERS, or to
manually add people to the relevant GitHub organizations, by just
providing an override key that works for all users. Of course, it is
the responsibility of the administrator to ensure that the override
key is cycled out regularly.

To that end, we introduce the concept of an administrator role, which
has permissions for all read/write/uploads (and, of course, to set the
upload override key). This requires some updates to the test suite as
the default adminstrator account is yours truly, which invalidates some
of the existing permission-setting tests.