Transition to a simplified API that only handles the uploads. (#12)
This is a complete redesign of the API for Bioconductor applications.  Namely,
it eliminates all GET endpoints, as the bucket is now assumed to be public and
directly accessible. We consider an extra level of hierarchy - namely
versioned assets within each project, which allows different parts of the
project to evolve at different rates. We distinguish between owners and
uploaders in the permissions, with scoped authorizations for individual
uploader accounts. Finally, we add a concept of probation whereby external
contributions can be reviewed by owners before approval/rejection.