Added an endpoint to refresh the latest version for an asset.
This is occasionally necessary if it gets out of sync, because the lock
isn't completely foolproof and we could get simultaneous updates to
..latest; so we just recompute the latest from summary timestamps.

We factor out the logic from the version removal endpoint into a
separate utility file and then re-use it in this new endpoint. Some
refactoring of the tests is also performed to re-use probation code.