Improve concurrency of endpoint handler with multiple async uploads.
This should improve the performance of these endpoints. It should be harmless
to have an undefined execution order for the uploads as nothing else in the
endpoint handler will depend on the uploads made by that endpoint.

In theory, we could do this with some of the fetch requests, but it gets too
confusing to read as the logic is distributed all over the handler function.