Respect limits on the number of keys for the R2 delete() method.
We only call the method if we have a non-zero number of keys; and for greater
than 1000 keys, we split the request into multiple chunks so that we don't hit
the per-request limit. This gives us a bit more breathing room for very large
projects, though eventually we'll be capped by the subrequest limit.